content_management/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.4
django-filter==23.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
Faker==19.3.1
httpie==3.2.2
idna==3.4
jwcrypto==1.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.2.0
oauthlib==3.2.2
pycparser==2.21
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0